requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
cycler==0.11.0
fonttools==4.29.1
html5lib==1.1
kiwisolver==1.3.2
lxml==4.7.1
matplotlib==3.5.1
MouseInfo==0.1.3
numpy==1.22.1
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
gym~=0.21.0